refactor : calculator에서 사용되는 inputFiled, tipCard, carbonEmissionCard를 각각의 List 컴포넌트안에서 제공하는 로직으로 변경
