+++ src/app/(features)/calculator/result/[year]/[month]/page.tsx
@@ -9,8 +9,8 @@ import { toJpeg } from "html-to-image";
 import Image from "next/image";
 import Loading from "@/components/calculator/Loading";
 import ThisMonthChart from "@/components/calculator/ThisMonthChart";
-import SectionCard from "@/components/calculator/SectionCard";
-import TipCard from "@/components/calculator/TipCard";
+import TipCardSection from "@/components/calculator/TipCardSection";
+import CarbonEmissionCardList from "@/components/calculator/CarbonEmissionCardList";
 
 const currentMonth = new Date().getMonth() + 1;
 const MIN_LOADING_TIME = 1000; // 최소 로딩 시간
@@ -80,24 +80,6 @@ const ResultPage: React.FC = () => {
   if (isLoading) {
     return <Loading />;
   }
-
-  // 최대 배출량 품목
-  const highestCo2Value = Math.max(
-    currentData?.electricity_co2 ?? 0,
-    currentData?.gas_co2 ?? 0,
-    currentData?.water_co2 ?? 0,
-    currentData?.waste_co2 ?? 0,
-    currentData?.car_co2 ?? 0
-  );
-
-  // 최저 배출량 품목
-  const lowestCo2Value = Math.min(
-    currentData?.electricity_co2 ?? 0,
-    currentData?.gas_co2 ?? 0,
-    currentData?.water_co2 ?? 0,
-    currentData?.waste_co2 ?? 0,
-    currentData?.car_co2 ?? 0
-  );
 
   return (
     <>
@@ -179,103 +161,24 @@ const ResultPage: React.FC = () => {
                     항목 별 탄소 배출량
                   </p>
                   <div className="flex flex-wrap w-full md:w-[1200px] gap-[12px] md:gap-[30px]">
-                    <SectionCard
-                      logo={"/calculate/electricity_color.svg"}
-                      title={"전기"}
-                      usageValue={currentData?.electricity_usage}
-                      co2Value={currentData?.electricity_co2}
-                      isHighest={
-                        currentData?.electricity_co2 === highestCo2Value
-                      }
-                      isLowest={currentData?.electricity_co2 === lowestCo2Value}
-                      unit="kwh"
-                    />
-                    <SectionCard
-                      logo={"/calculate/gas_color.svg"}
-                      title={"가스"}
-                      usageValue={currentData?.gas_usage}
-                      co2Value={currentData?.gas_co2}
-                      isHighest={currentData?.gas_co2 === highestCo2Value}
-                      isLowest={currentData?.gas_co2 === lowestCo2Value}
-                      unit="m³"
-                    />
-                    <SectionCard
-                      logo={"/calculate/water_color.svg"}
-                      title={"수도"}
-                      usageValue={currentData?.water_usage}
-                      co2Value={currentData?.water_co2}
-                      isHighest={currentData?.water_co2 === highestCo2Value}
-                      isLowest={currentData?.water_co2 === lowestCo2Value}
-                      unit="m³"
-                    />
-                    <SectionCard
-                      logo={"/calculate/car_color.svg"}
-                      title={"교통"}
-                      usageValue={currentData?.car_usage}
-                      co2Value={currentData?.car_co2}
-                      isHighest={currentData?.car_co2 === highestCo2Value}
-                      isLowest={currentData?.car_co2 === lowestCo2Value}
-                      unit="km"
-                    />
-                    <SectionCard
-                      logo={"/calculate/waste_color.svg"}
-                      title={"폐기물"}
-                      usageValue={currentData?.waste_volume}
-                      co2Value={currentData?.waste_co2}
-                      isHighest={currentData?.waste_co2 === highestCo2Value}
-                      isLowest={currentData?.waste_co2 === lowestCo2Value}
-                      unit="Kg"
-                    />
+                    <CarbonEmissionCardList currentData={currentData} />
                   </div>
                 </div>
               </div>
             </div>
 
+            {/* 세번째 섹션 데이터 제공 */}
             <div>
               <div>
                 <div className="text-[14px] md:text-[24px] font-semibold mb-[32px]">
                   일상 속 에너지 절약법
                 </div>
-                <div className="flex flex-col gap-3 mb-[48px]">
-                  <TipCard
-                    tipLogo={"/calculate/electricity_white.svg"}
-                    tipTitle={"LED 조명으로 교체하기"}
-                    tipContent={
-                      "LED는 기존 전구보다 전력 소모가 적어 전기 절약 효과가 큽니다."
-                    }
-                  />
-                  <TipCard
-                    tipLogo={"/calculate/gas_white.svg"}
-                    tipTitle={"겨울철 실내 온도 낮추기"}
-                    tipContent={
-                      "난방 온도를 조금만 낮추고,담요나 양말을 활용해 따뜻함을 유지해 보세요."
-                    }
-                  />
-                  <TipCard
-                    tipLogo={"/calculate/water_white.svg"}
-                    tipTitle={"세탁물 모아서 한 번에 세탁하기"}
-                    tipContent={
-                      "세탁기물을 한 번에 모아서 사용해 물과 전기를 함께 절약하세요."
-                    }
-                  />
-                  <TipCard
-                    tipLogo={"/calculate/car_white.svg"}
-                    tipTitle={"출퇴근 카풀하기"}
-                    tipContent={
-                      "같은 방향으로 가는 사람들과 차를 함께 타면 연료와 교통비를 아낄 수 있어요."
-                    }
-                  />
-                  <TipCard
-                    tipLogo={"/calculate/waste_white.svg"}
-                    tipTitle={"텀블러와 에코백 사용하기"}
-                    tipContent={
-                      "일회용 컵과 비닐봉투 대신 텀블러와 에코백을 사용해 폐기물을 줄여보세요."
-                    }
-                  />
-                </div>
+                <TipCardSection />
               </div>
             </div>
           </div>
+
+          {/* 이미지 다운로드 버튼 */}
           <div className="flex justify-center">
             <button
               className="w-[320px] md:w-[360px] h-[60px] px-8 bg-[#0D9C36] text-white rounded-[85px] text-[18px] font-semibold border-none"
